backend/requirements.txt
asgiref==3.6.0
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
decorator==4.4.2
Django==4.2
django-cors-headers==3.14.0
django-debug-toolbar==4.0.0
djangorestframework==3.14.0
ffmpeg-python==0.2.0
filelock==3.12.0
fsspec==2023.5.0
future==0.18.3
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.18.0
google-auth-httplib2==0.1.0
google-cloud==0.34.0
google-cloud-speech==2.19.0
google-cloud-vision==3.4.1
googleapis-common-protos==1.59.0
grpcio==1.54.2
grpcio-status==1.54.2
gTTS==2.3.2
httplib2==0.22.0
huggingface-hub==0.14.1
idna==3.4
imageio==2.28.1
imageio-ffmpeg==0.4.8
install==1.3.5
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.2
moviepy==1.0.3
mpmath==1.3.0
networkx==3.1
nltk==3.8.1
numpy==1.24.3
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
proglog==0.1.10
proto-plus==1.22.2
protobuf==4.23.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
regex==2023.5.5
requests==2.30.0
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
SpeechRecognition==3.10.0
sqlparse==0.4.4
sympy==1.12
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.29.2
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
